refactor test_runner module
